[frontend] rename directory and fix deepscan (#1633)
